pubsub-data-sync/requirements.txt
functions-framework==3.*
google-cloud-pubsub==2.*
google-cloud-bigquery==3.*